service: move accountid, credential, locker, network to internal
