+++ service/internal/network/error.go
