+++ service/internal/network/network.go
