+++ service/internal/network/splitter.go
