+++ service/internal/network/splitter_test.go
